requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
databases[aiosqlite]
python-dotenv
pydantic-settings
python-jose
python-multipart
passlib[bcrypt]